refactor(ts-sdk): instruction module
