+++ ts-sdk/examples/ucs03.ts
@@ -1,5 +1,5 @@
 import { toHex } from "viem"
-import { Batch, encodeAbi, Forward, FungibleAssetOrder, Multiplex } from "../src/ucs03/index.js"
+import { Batch, encodeAbi, Forward, FungibleAssetOrder, Multiplex } from "../src/ucs03/instruction.js"
 import { forwardAbi } from "../src/evm/abi/index.js"
 
 export const exampleBatchInstruction = Batch([
